added citygraph

+++ src/GRAPH/Graph.java
@@ -6,7 +6,7 @@ import java.util.List;
 import java.util.Map;
 
 public class Graph {
-    HashMap<Vertex, HashMap<Integer, ArrayList<Vertex>>> adj_list;
+    HashMap<Vertex, HashMap<Double, ArrayList<Vertex>>> adj_list;
 
     public Graph() {
         this.adj_list = new HashMap<>();
@@ -26,11 +26,11 @@ public class Graph {
         ArrayList<Vertex> eV2 = adj_list.get(v2).get(0);
         if (eV1 == null) {
             eV1 = new ArrayList<>();
-            adj_list.get(v1).put(0, eV1);
+            adj_list.get(v1).put((double)0, eV1);
         }
         if (eV2 == null) {
             eV2 = new ArrayList<>();
-            adj_list.get(v2).put(0, eV2);
+            adj_list.get(v2).put((double)0, eV2);
         }
         eV1.add(v2);
         eV2.add(v1);
@@ -50,11 +50,11 @@ public class Graph {
     @Override
     public String toString() {
         StringBuilder sb = new StringBuilder();
-        for (Map.Entry<Vertex, HashMap<Integer, ArrayList<Vertex>>> entry : adj_list.entrySet()) {
+        for (Map.Entry<Vertex, HashMap<Double, ArrayList<Vertex>>> entry : adj_list.entrySet()) {
             Vertex v = entry.getKey();
             sb.append(v).append(" -> ");
-            HashMap<Integer, ArrayList<Vertex>> edges = entry.getValue();
-            for (Map.Entry<Integer, ArrayList<Vertex>> entry2 : edges.entrySet()) {
+            HashMap<Double, ArrayList<Vertex>> edges = entry.getValue();
+            for (Map.Entry<Double, ArrayList<Vertex>> entry2 : edges.entrySet()) {
                 ArrayList<Vertex> e = entry2.getValue();
                 for (Vertex v2 : e) {
                     sb.append(v2).append(", ");
